(maint) Devise auth now works
Added bootstrap elements to registration and login pages. Added a
warning bar when you either log in or out.

Problem discovered with TBS 3.0:

If you try to use the example log in page that TBS provides, Devise will
not work. That's because the second html element that they had was a
form tag. This form overwrote Devise's form, and it broke auth. Changing
it to a div tag fixed this problem and kept the css elements. Woo!