(maint) Improve home page for user and change alert
This commit adds a few visual changes to the dashboard home page
post and pre login. Instead of showing links on the nav-bar, this
commit adds buttons with colors to make it more visually appealing. When
logged in, it will show users a user table (mostly for debugging
right now, will remove later). It will also show a dropdown on the
users profile in the navbar. Right now it only links to editing the
profile, but later it might have more features.

Another change was to fix the alert for logging in and out from
"warning" to "info" The blue is a better color for this action, since
logging in and out don't need to be warnings.