(#6 #7) Add JBuilder and Demo REST API
Add Jbuilder to project and present a demo REST API that gives a list of
users in json format.

Later, in the next iteration, this REST API will
give the user access to the MTD model (auth done by devise tokens).
Likely to be initiated by Angular or through command line.