(#23) Add example button dropdown
In this commit I move the text editor box into a frame, and place the
dropdown in a frame next to the text editor. Add custom css styling for
lists in AngularJS. Before it was just for li elements, but the custom
styling messed up other dropdowns, and since it was only meant for the
angularjs demo, I added a custom style to get rid of that.