(maint) Add d3 demo graphs
This commit adds two main graph demos to the dashboard. The first demo
on the home page shows a graph over time for services the dashboard
could be using. The other demo is a config overlook that shows the state
of the system depending on what the fullstack.yaml file has been
configured to by the user.