Remove unneeded dupe js functions
Prior to this commit, I had duplicate js files just to get something
working on the animated bar chart demo. This commit removes those dupe
js files and modifies the original js file to accept parameters to
customize the animated bar charts. It now takes a div element to tell
the function where the D3 will be applied to, and the specific css
element it will use for bar color.