Add beginnings of network view
This commit adds the start of the network view. It adds an example D3
graph that will map what the network looks like depending on what the
fullstack.yaml file has. It will also have a network table that details
the health of the services/clusters on the system. It adds a dashboard
helper function that reads in the yaml file and extracts the service
names needed to create the table and other graphs.