Add file selector and dropdown config
This commit adds a modal file selector that will be able to allow users
to select a file they want to load into the ace text editor. Currenty it
just opens a file dialog and selects the file from the users machine.
Next step is to load that file into the ace editor.

Dropdown config is to show some automation from the fullstack file.
Eventually they will be able to select and config MTD through dropdowns
and checkboxes similar to openstack.