(#33) Add notification after successful file upload
This commit adds a notification to the web page if a user uploads a
file. Prior to this commit the webpage would just reload as if nothing
happened, making it confusing on if it worked or not. This now displays
a simple notification (similar to the login ones) saying that the file
was uploaded.