(#35) Add more robust User table demo
This commit adds a sample User table for the admin panel view. It adds
checkboxes in case you want to perform multiple operations on many
users, it adds an emailto link so when you select their email address it
will open their mail client, it displays if the user is enabled or not,
and instead of deleting a user you can disable users (although for now
these are only just buttons).