(#32) Add auto-loading forms for user configuration
This commit adds the beginnings of the auto-configuration feature of the
dashboard. A user types in a role name, and then automatically another
form will appear for a user to add in a sub-goal name. Then after they
have enetered that name in, they can begin to add the role list.