Add input validation and config process
This commit adds a couple features for the manage page:

First it has a simple and extendable validation for user input. Simply,
right now this just makes sure the user doesn't enter in something blank
or too long. In the future, more requirements might be needed within the
function.

Second, this adds a more complex process of defining the goals config.
It walks the user through what to configure, and when they have finished
defining everything, it asks if the user is ready to continue
configuration based on what they had defined.