Move text editor button location in code
Prior to this commit, the text editor buttons were on a different row.
This would pose as a problem since I am making a different kind of
configuration on the right side. The buttons would be on the row below,
making them further away from the editor as more HTML elements are added
to the right column. So this commit moves those buttons to the same row
as the text editor.