(#38) Generate instance information based on json file
This commit adds some logic to grab a json file, parse it, place the
data in a hash, and send it to the controller. The controller makes that
hash available for the view so it can traverse the hash and generate a
table from the returned hash. Currently this is static from a file
within rails, however in the future this data will be gathered through a
REST API, and this data will be associated with a given instance on the
view portion of this page. It will only be displayed if someone clicks
that instance.